lease decoupled from storage
